+++ pycore/src/myarm_sdk/adapters/robot_arm/__init__.py
@@ -1,0 +1,6 @@
+"""Robot-arm adapters."""
+
+from .fake_robot_arm import FakeRobotArm
+from .pymycobot_robot_arm import PymycobotRobotArm
+
+__all__ = ["FakeRobotArm", "PymycobotRobotArm"]